Fixes falsey test for required input check
`$codeShowInvisibles` and `$codeUseSoftTabs` were not using a strict test to check against having valid data (which in this case could be 0 [zero] for false). Adding `isset` to check if the value is not empty/null so that a zero value does not cause a failure in submission when setting the false value provided in the configuration form for these two options. (NOTE: Tab size theoretically should allow a zero value as well, but what monster would do such a thing?? I left that logic error in place for convention. 😉)